+++ packages/shared/prisma/schema.prisma
@@ -551,16 +551,18 @@ model CronJobs {
 }
 
 model Dataset {
-  id           String        @default(cuid())
-  projectId    String        @map("project_id")
-  name         String
-  description  String?
-  metadata     Json?
-  project      Project       @relation(fields: [projectId], references: [id], onDelete: Cascade)
-  createdAt    DateTime      @default(now()) @map("created_at")
-  updatedAt    DateTime      @default(now()) @updatedAt @map("updated_at")
-  datasetItems DatasetItem[]
-  datasetRuns  DatasetRuns[]
+  id             String        @default(cuid())
+  projectId      String        @map("project_id")
+  name           String
+  description    String?
+  metadata       Json?
+  webhookUrl     String?       @map("webhook_url")
+  webhookPayload Json?         @map("webhook_payload")
+  project        Project       @relation(fields: [projectId], references: [id], onDelete: Cascade)
+  createdAt      DateTime      @default(now()) @map("created_at")
+  updatedAt      DateTime      @default(now()) @updatedAt @map("updated_at")
+  datasetItems   DatasetItem[]
+  datasetRuns    DatasetRuns[]
 
   @@id([id, projectId])
   @@unique([projectId, name])
